backend/requirements.txt
fastapi
uvicorn
chromadb
sentence-transformers
pandas
python-dotenv
httpx
groq
